ramips: mt7621: migrate ledtrig-usbdev to ledtrig-usbport
This patch set a single LED to be triggered by both USB 1/2 and the
USB 3 port.

Do it via userspace for now, till the correct devicetree syntax of
doing is figured out.

This patch has the side effect of preventing LuCI from erroneously
deleting the trigger due to the fact that usbdev is gone and is only
emulated now.

Signed-off-by: Daniel Gimpelevich <daniel@gimpelevich.san-francisco.ca.us>
Signed-off-by: Mathias Kresin <dev@kresin.me>